moved styled elements into new directory 'emotion' nested uncer 'components'
